Use string tempids
We used to hash strings and use the number as a tempid. This was stable in a single transaction but had a risk of collision.

We now use a recent feature of Datomic, to use strings as tempids. This is easier to debug when reading transaction data.

Since we use the same strings as :uri with unique/identity, subsequent transactions with the same class names unify with the existing entities. That lets us make a very tiny change to take advantage of this feature.